+++ example/next.config.mjs
@@ -1,4 +1,9 @@
+import path from 'path'
+
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+  outputFileTracingRoot: path.join(process.cwd(), '..'),
+  transpilePackages: ['@reclaimprotocol/js-sdk'],
+};
 
 export default nextConfig;